move fonts to `/include/fonts`
